day 2(part 1): gift shop

+++ src/day_1.rs
@@ -5,7 +5,7 @@ use crate::Solution;
 pub struct Day1;
 
 impl Solution for Day1 {
-    fn part1() -> u32 {
+    fn part1() -> u64 {
         let input = Path::new("input/day1.txt");
         let contents = fs::read_to_string(input).expect("Something wrong with opening this file");
         let mut cur_iter = 50_i32;
@@ -27,11 +27,11 @@ impl Solution for Day1 {
         total
     }
 
-    fn part2() -> u32 {
+    fn part2() -> u64 {
         let input = Path::new("input/day1.txt");
         let contents = fs::read_to_string(input).expect("Something wrong with opening this file");
         let mut cur_iter = 50_i32;
-        let mut total = 0_u32;
+        let mut total = 0_u64;
         contents.lines().for_each(|line| {
             let (cmd, num_str) = line.split_at(1);
             let num = num_str.parse::<i32>().unwrap();
@@ -48,7 +48,7 @@ impl Solution for Day1 {
 
             if cur_iter != 0 {
                 let excess = (cur_iter.unsigned_abs() - 1) / 100;
-                total += excess;
+                total += excess as u64;
             }
 
             cur_iter = cur_iter.rem_euclid(100);
