+++ src/solution.rs
@@ -1,4 +1,4 @@
 pub trait Solution {
-    fn part1() -> u32;
-    fn part2() -> u32;
+    fn part1() -> u64;
+    fn part2() -> u64;
 }
